signing/requirements.txt
selenium==4.1.0
python-dotenv==0.10.3
pandas==1.3.4